Refactor : 루틴 관련 엔티티 하나의 패키지로 보관
